Fix: customAxios 세션 만료 에러 예외 처리 조건 추가

+++ shared/common/utils/customAxios.ts
@@ -12,7 +12,11 @@ const customAxios = axios.create({
 customAxios.interceptors.response.use(
   (response) => response,
   async (error: AxiosError<ApiResponse>) => {
-    if (isAxiosError(error) && error.response?.status === 401) {
+    if (
+      isAxiosError(error) &&
+      error.response?.status === 401 &&
+      error.response?.data.message === "만료된 세션이에요."
+    ) {
       if (typeof window !== "undefined" && process.env.TEST_ENV !== "SSR") {
         await fetch("/api/auth/session-cookie", { method: "DELETE" });
         redirect("/session-expired");
